refactor(executors): extract pure payload construction from claude-web (1056→835)
Port of upstream diegosouzapw/OmniRoute#6006. Extracts pure payload
builders/transforms into open-sse/executors/claude-web/payload.ts, leaving
the host executor to import them back. Adds a split-boundary unit test.
